+++ front-end/package.json
@@ -10,11 +10,15 @@
   },
   "dependencies": {
     "@clerk/clerk-expo": "^0.18.17",
+    "@react-navigation/bottom-tabs": "^6.5.8",
+    "@react-navigation/native": "^6.1.7",
     "expo": "~49.0.7",
     "expo-status-bar": "~1.6.0",
     "react": "18.2.0",
     "react-native": "0.72.3",
-    "react-native-paper": "^5.10.1"
+    "react-native-paper": "^5.10.1",
+    "react-native-safe-area-context": "^4.7.1",
+    "react-native-screens": "^3.24.0"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0"
